Update flexpanel to upstream phosphor changes
